initial testing cases

+++ matrix.c
@@ -29,7 +29,7 @@ struct matrix* matrix_new(int n_row, int n_col) {
     return new_matrix;
 }
 
-struct matrix* matrix_from_array(double* data, int n_row, int n_col) {
+struct matrix* matrix_from_array(float* data, int n_row, int n_col) {
     assert(n_row >= 1 && n_col >= 1);
     struct matrix* M = matrix_new(n_row, n_col);
     for(int i = 0; i < n_row * n_col; i++) {
